requirements.txt
torch>=2.1.0
transformers>=4.30.0
python-docx>=0.8.11
PyPDF2>=3.0.1
fastapi>=0.95.0
flask>=2.2.0
numpy>=1.23.0